requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
pandas
reportlab
python-magic==0.4.25
